webkit: page contents not shown correctly on changing to landscape mode
The page content is not displayed properly because the most recent
pictureset maintained in the BaseLayerAndroid is not painted.As we know
the TreeManger maintains 3 layers i.e., drawing , painting , queue.
These layers were shifted painting to drawing and queued to painting as
the BaseTiles gets painted by paint operation in TextureGeneration
thread. The layer swap did not happens due to mismatch in zoom state.
The zoom manager has to process the new zoom scale request and updates
its state accordingly. In the scenario, the previously requested zoom
state is changed to the current state before it gets handled. As a
result the new request is never scheduled and results in layers not
getting swapped until new zoom scale request happens. Zoom scale can
changes due to orientation changes.
This patch ensure the zoom scale request is schedule in the scenario
explained here.

Change-Id: I18c9a060f5587e7709147c54a6c002865753a8ec